🔨 fasttrack: deprecate old cli (#2372)
