+++ apps/README.md
@@ -1,7 +1,3 @@
 ⚠️ WORKING ON IMPROVING THIS DOCS
 
-This directory contains apps built around and for the ETL.
-
-
-- [`wizard`](./wizard) - Tool to generate ETL steps. It is a great starting point when creating new steps.
-- [`backport`](./backport) - Tool to backport datasets from Grapher.
+This directory contains apps built around and for the ETL. We are slowly integrating all our apps into Wizard, but some of their code still lives outside of it.
