--- apps/fasttrack/styles.css
@@ -1,27 +1,0 @@
-table.diff {
-  font-family: Courier;
-  border: medium;
-  font-size: 12px;
-}
-table.diff td {
-  padding-top: 2px;
-  padding-bottom: 2px;
-}
-.diff_header {
-  background-color: #e0e0e0;
-}
-td.diff_header {
-  text-align: right;
-}
-.diff_next {
-  background-color: #c0c0c0;
-}
-.diff_add {
-  background-color: #aaffaa;
-}
-.diff_chg {
-  background-color: #ffff77;
-}
-.diff_sub {
-  background-color: #ffaaaa;
-}
