+++ pyproject.toml
@@ -10,7 +10,6 @@ etlcli = 'apps.cli:cli'
 etlwiz = 'apps.wizard.cli:cli'
 etlr = 'etl.command:main_cli'
 etl-wizard = 'apps.wizard.cli:cli'
-fasttrack = 'apps.fasttrack.cli:cli'
 compare = 'etl.compare:cli'
 backport = 'apps.backport.backport:backport_cli'
 
@@ -97,7 +96,6 @@ plotly = ">=5.7.0"
 cookiecutter = ">=2.1.1"
 types-PyYAML = ">=6.0.12"
 hydra-core = ">=1.3.0"
-pywebio = ">=1.8.2"
 boto3-stubs = { extras = ["s3"], version = ">=1.21.16" }
 gspread = "^5.10.0"
 # unpinning those would introduce tons of type errors
